Renamed main monitoring script and added support for all 2110 and 2022 products
